# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus CRM API",
    "version": "1.0.0",
    "description": "REST API for the Nexus CRM application.\n\n## Authentication\n\nTwo auth methods are supported:\n\n- **BearerAuth** \u2013 `Authorization: Bearer <token>`. Per-user API token generated in the dashboard. Scoped to the token owner's data (admin tokens get cross-tenant read access).\n- **CookieAuth** \u2013 Session cookie set by the Better Auth `/api/auth` flow (Google OAuth).\n\nToken management (POST/DELETE `/api/token`) requires a session cookie. File downloads require CookieAuth or BearerAuth; unauthenticated sharing uses per-link `/s/{code}` URLs scoped to the link owner."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-user API token generated via the dashboard (POST /api/token). Scoped to the token owner's data. Admin users' tokens get cross-tenant read access."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by the Better Auth flow at `/api/auth`. Admin users (isAdmin=true) bypass per-user filtering."
      }
    },
    "schemas": {
...
      "ApplicationRef": {
        "type": "object",
        "description": "Minimal application reference (used inside Document).",
        "required": ["id", "company", "role"],
        "properties": {
          "id":      { "type": "integer", "example": 42 },
          "company": { "type": "string",  "example": "Acme Corp" },
          "role":    { "type": "string",  "example": "Senior Engineer" }
        }
      },
...
      "Document": {
        "type": "object",
        "description": "An uploaded document (PDF, image) attached to one or more applications.",
        "required": ["id", "userId", "filename", "originalName", "size", "mimeType", "uploadedAt", "applications"],
        "properties": {
          "id":           { "type": "integer", "example": 7 },
          "userId":       { "type": "string",  "example": "user_abc123" },
          "filename":     { "type": "string",  "example": "3f2504e0-4f89-11d3-9a0c-0305e82c3301.pdf", "description": "UUID-based filename on disk." },
          "originalName": { "type": "string",  "example": "my-resume.pdf" },
          "size":         { "type": "integer", "example": 204800, "description": "File size in bytes." },
          "mimeType":     { "type": "string",  "example": "application/pdf", "enum": ["application/pdf", "image/jpeg", "image/png", "image/webp"] },
          "documentType": { "type": "string" },
          "state": { "type": "string", "enum": ["draft", "current", "submitted", "superseded", "historical", "orphaned"] },
          "version": { "type": "integer", "minimum": 1 },
          "contentHash": { "type": ["string", "null"] },
          "source": { "type": ["string", "null"] },
          "generatedAt": { "type": ["string", "null"], "format": "date-time" },
          "submittedAt": { "type": ["string", "null"], "format": "date-time" },
          "submissionId": { "type": ["string", "null"] },
          "uploadedAt":   { "type": "string",  "format": "date-time" },
          "applications": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApplicationRef" }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions — admin privileges required or resource does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Session required" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found or does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Not found" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid or missing request parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/documents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer", "minimum": 1 },
          "description": "Document ID."
        }
      ],
      "patch": {
        "operationId": "updateDocumentLinks",
        "summary": "Update document application links",
        "description": "Replaces the set of applications linked to this document. Provide a complete `applicationIds` array — this is a *set* operation, not a merge. **Requires a session.**",
        "tags": ["Documents"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["applicationIds"],
                "properties": {
                  "applicationIds": {
                    "type": "array",
                    "items": { "type": "integer" },
                    "description": "Complete list of application IDs to link. Only IDs owned by the session user are accepted.",
                    "example": [42, 43]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated document with new application links.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Document" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteDocument",
        "summary": "Delete document",
        "description": "Deletes the document record and removes the file from disk. **Requires a session.**",
        "tags": ["Documents"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "responses": {
          "204": {
            "description": "Document deleted. No response body."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
...
}